Preparing for editing and formatting
The only change worth of mention is the .gitignore file, which
now makes git ignore files that clangd and clang-format use
(eg. compile_commands.json, if it happens to be located outside
build directory, and .cache)

All other changes are due to file formatting and *will be
changed* due to them being inconsistant etc.

Note to self: make sure to copy the .clang-format file to
project root directory.