Fixed lexer and symbol analysis
The "aware" connection was used incorrectly, so the context-free grammar file ./policy/docs/cfg.txt has been fixed.
Consequently, AWARE enum entry in `Term` and `Nont` enums was being used wrong, which was also fixed now.

In policy.cc, every `std::` was removed and `using namespace std` was added in the beginning section of the file.

Further work was done on the implementation of the "aware" connections as well. See synan.cc and ast.cc.