Progress in Abstract Syntax Tree implementation This is a temporary commit, which is going to undergo testing.
The main changes were that a few classes were added to header files (for
example `ast_aware_t`) which are made according to already existing
classes.

There was a lot of documentation and explanations added (comments)
which help with understanding what the code does. For now, this is also
the place where TO-DOs are written down (use `grep` to search for
"TODO")

For easier navigating and understanding of changes, all modified/added
files are listed bellow.

	modified:   parser/tag_parser.cc
	modified:   policy/ast.cc
	modified:   policy/ast.h
	modified:   policy/policy.cc
	modified:   policy/policy.h
	modified:   policy/synan.cc
	modified:   policy/synan.h
	new file:   scripts/install-sh
	new file:   tests/hello-privacy/hello-privacy
	modified:   tests/hello-privacy/string-public/hello-privacy.policy

# Conflicts:
#	parser/tag_parser.cc
#	policy/policy.cc