Constructor for policy_t now processes aware files
Or so it may seem. This commit will undergo further testing. For now,
running the program with hello-privacy's string-public successfuly
exits, although setting up the `aware_connections` property of policy_t
is the only thing being achieved for now.

Note: for now, only basic topology is supported.

Note: the implementation is a bit cumbersome. See the comments in
policy.cc.