WORKING AWARE TAG PARSING for basic topologies
The tag-parser now correctly outputs a file which now contains the
number of aware topologies (besides the number of topologies and
perimeter-guards). It comes last:

    [num. of topologies] [num. of pgs] [num. of awares]

Aware connections are specified as such:

    [name] [num. of pairs] [source1] [end1] [source2]...

NOTE: for now, ONLY BASIC TOPOLOGY TYPE CAN BE USED IN AWARE POLICY.
Linear and Expression types will be added eventually if needed.

Another directory named `string-public-aware` was created inside the
`tests/hello-privacy` directory, which is (for now) just a copy of
`string-public` directory, except for the .policy file, which was
modified appropriately.